backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
google-genai
groq
requests
authres
dkimpy
dnspython
python-dotenv
python-whois
# Threat intelligence & attachment scanning
stix2
yara-python
oletools
pymisp
